+++ src/main/java/com/auth/model/Product.java
@@ -2,6 +2,7 @@ package com.auth.model;
 
 
 import jakarta.persistence.Entity;
+
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -14,7 +15,7 @@ public class Product {
    private String name;
    private String category;
    private String status;
-   private int price;
+   private Double price;
    private String images;
    public Product() {
 	super();
@@ -44,10 +45,10 @@ public class Product {
    public void setStatus(String status) {
 	this.status = status;
    }
-   public int getPrice() {
+   public Double getPrice() {
 	return price;
    }
-   public void setPrice(int price) {
+   public void setPrice(Double price) {
 	this.price = price;
    }
    public String getImages() {
